Avoid `url_for` for anchor links
Rails' `url_for` method (which object_path) uses is overkill for simply
creating a link to an anchor on the current page.

Each call to `object_path` allocates a Hash, and then the Rails
`url_for` logic performs a lot more allocations and router lookups,
just to inevitably give us `/2.7/o/string#method-i-chr` for the
`/2.7/o/string` page. For an anchor, the `/2.7/o/string` is actually
irrelevant. `#method-i-chr` as the href will work just fine.

Additionally, the method names are CGI escaped, and since a single page
at least 2 or 4 times (depending on if there's both an instance and
class variable with the same name), we can remove great deal of
allocations by caching the escaped name for the duration of the request.

Improvement -19k allocations

Before:
Completed 200 OK in 54ms (Views: 40.2ms | Allocations: 44923)

After:
Completed 200 OK in 38ms (Views: 24.8ms | Allocations: 26282)