Cache escaped method names
Every time objects#show is loaded, all of the class and instance
methods are escaped to be used as an anchor link to the method
on the current page.

The thing is, the value a single method name escapes to is constant,
so there's no need to calculate it every request. Furthermore,
there's a lot of overlap between classes on method names
(how many classes have a `new` method, for example?).

By storing a cache of the html anchors for a method, we can reduce
a significant amount of allocations on the objects#show page.

For `/2.7/o/string`, this removes about 7,000 allocations.

Before:
Completed 200 OK in 51ms (Views: 38.9ms | Allocations: 26304)

After:
Completed 200 OK in 39ms (Views: 25.9ms | Allocations: 24993)

Removes ~1,300 allocations