DRY the header
The header was copy/pasted between the home page and header layout
because of a few intentional differences for the home page:
* doesn't have a fixed top header
* doesn't have a search button in the header
* includes GitHub link

There was one difference that I believe was unintentional,
some spacing on the left/right and top/bottom was bigger on the home
page than the rest of the site, which made the header jump around
as you navigated between pages. I made the home page be the same
as the rest of the site.

There was one difference that I wasn't sure if it was intentional,
but I kept it. The version number in the header on mobile is bigger
on the home page than the rest of the site. Given that there's no
search bar, there's a little more space, so I thought this was
intentional and kept it.