Make RDoc links be consistent with the rest of the Rails app
RDoc's links are relative paths from the current file,
where the path is the exact casing as the class, and end in .html.

These links work fine from the Objects#show page, but not on
search results.
Additionally, they'll result in extra page caching, as the URL
differs from what the app generates.

We can fix this by parsing the URLs, and having the router
generate us a proper URL.

Fixes #361