Fix FontAwesome icon flicker
The font awesome library uses Javascript to convert <i> tags
into inlined SVGs. Because the entire application pack is deferred,
this javascript does not run until much later, causing the icons to
pop into view with every page load.

By separating the icons out of the main JS file we can load the
icons immediately, independently of the rest of the site's JS.
(The rest of the JS is just progressive enhancement, so we can wait
for the HTML to render before processing it.)