Switch to using `keydown` event
If we use `keyup` we won't be able to call `event.preventDefault()` to
stop the default behavior of whatever key was pressed. This is important
because we need to override the Enter key's behavior to select the
currently highlighted suggestion instead of searching with the current
contents of the search field. Also, we want the Up & Down keys to
navigate the search suggestions without moving the text cursor to the
beginning/end of the text field.