Add `onKeyup` action for non-special keys
We switched from listening for `keyup` events to `keydown` events in
832b98133b56a00fd940f1c142a15bb0438cab62 because we needed to be able to
call `preventDefault` when the event was an arrow up, arrow down, or
Enter/Return key press. Using `keyup` would mean that we would call
`preventDefault` too late to stop the default behavior.

However, this introduced a bug where the query that we send to the
server is missing its last character. At the time that the `keydown`
action is executed, the input field doesn't yet contain the character
that will be typed when the key goes up.

The easiest way I know of to solve this is to register functions for
both `keyup` and `keydown`. We can handle special characters in
`onKeydown` and normal characters in `onKeyup`.