Fix duplicate path fragments from PathCleaner.clean
The PathCleaner generates path fragments according to the ruby
constants. This works by splitting the uri path provided by ruby into
class parts. When handling nested constants, the default class part list
contains all constant fragments from the constant name except the last one.

We need to do this because the ruby docs sometimes do not include the
parent class parts in the uri path. However sometimes they do include
those parent class parts.

This led to class parts being duplicated in the cleaned path. One
example of this is `Thread::Queue#deq`. The alias method path provided
by ruby is "Thread/Queue" and the constant is "Thread::Queue". When
cleaning the path we default to `['Thread']` and join "Thread" and
"Queue" afterwards, resulting in the path `thread/thread/queue` which
is not valid.

This patch tries to fix this by only adding new class parts if the
new part does not already exist in the list of class parts.

Fixes #1748